# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: VPN Admin API
  description: Control plane for commercial VPN (AmneziaWG).
  version: 0.1.0-rc.1
paths:
...
  /api/v1/peers:
    get:
      tags:
      - peers
      summary: List Peers
      description: 'List peers (devices). peer_id = device.id, node_id = device.server_id.

        Create peer: use POST /api/v1/devices/issue (with server_id or omit for load-balanced).

        Remove peer: use POST /api/v1/devices/{id}/revoke or /reset.'
      operationId: list_peers_api_v1_peers_get
      security:
      - HTTPBearer: []
      parameters:
      - name: node_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by node (server) id
          title: Node Id
        description: Filter by node (server) id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: active | revoked
          title: Status
        description: active | revoked
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerListOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/peers/{peer_id}:
    get:
      tags:
      - peers
      summary: Get Peer
      description: Get one peer (device) by id.
      operationId: get_peer_api_v1_peers__peer_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: peer_id
        in: path
        required: true
        schema:
          type: string
          title: Peer Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerListItemOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/peers/{peer_id}/migrate:
    post:
      tags:
      - peers
      summary: Migrate Peer Endpoint
      description: Move peer to another node. Remove from current, add on target,
        update device.server_id (spec 4.2).
      operationId: migrate_peer_endpoint_api_v1_peers__peer_id__migrate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: peer_id
        in: path
        required: true
        schema:
          type: string
          title: Peer Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MigrateBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/wg/peer:
    post:
      tags:
      - wg
      summary: Create Wg Peer
      description: 'Provision a VPN peer (spec: POST /wg/peer). Delegates to issue_service
        with load-balance when server_id omitted.'
      operationId: create_wg_peer_api_v1_wg_peer_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WgPeerCreateBody'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
      - HTTPBearer: []
  /api/v1/wg/peer/{pubkey}:
    delete:
      tags:
      - wg
      summary: Delete Wg Peer
      description: 'Remove peer by public key (spec: DELETE /wg/peer/{pubkey}). Revokes
        device and removes peer from node.'
      operationId: delete_wg_peer_api_v1_wg_peer__pubkey__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: pubkey
        in: path
        required: true
        schema:
          type: string
          title: Pubkey
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/servers/{server_id}/peers:
    post:
      tags:
      - servers
      summary: Create Server Peer
      description: Create a peer on the chosen server; return one-time config_download_url
        and qr_payload.
      operationId: create_server_peer_api_v1_servers__server_id__peers_post
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminIssuePeerRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminIssuePeerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
    get:
      tags:
      - servers
      summary: Get Server Peers
      description: Fetch peers from node with pubkey, handshake, rx/tx, status. Returns
        empty when node unreachable.
      operationId: get_server_peers_api_v1_servers__server_id__peers_get
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerPeersOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/agent/v1/peers:
    get:
      tags:
      - agent
      summary: Agent V1 Peers
      operationId: agent_v1_peers_api_v1_agent_v1_peers_get
      parameters:
      - name: server_id
        in: query
        required: true
        schema:
          type: string
          description: Server ID this agent manages
          title: Server Id
        description: Server ID this agent manages
      - name: X-Agent-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentV1PeerOut'
                title: Response Agent V1 Peers Api V1 Agent V1 Peers Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
      - AgentToken: []
      x-mtls-required: true
...
components:
  schemas:
...
    AdminIssuePeerPeerOut:
      properties:
        id:
          type: string
          title: Id
        server_id:
          type: string
          title: Server Id
        device_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Name
        public_key:
          type: string
          title: Public Key
        issued_at:
          type: string
          format: date-time
          title: Issued At
      type: object
      required:
      - id
      - server_id
      - device_name
      - public_key
      - issued_at
      title: AdminIssuePeerPeerOut
    AdminIssuePeerRequest:
      properties:
        user_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: User Id
        subscription_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Subscription Id
        label:
          anyOf:
          - type: string
          - type: 'null'
          title: Label
        device_limit:
          anyOf:
          - type: integer
          - type: 'null'
          title: Device Limit
        expires_in_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Expires In Days
        client_request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Request Id
        client_endpoint:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Endpoint
      type: object
      title: AdminIssuePeerRequest
    AdminIssuePeerResponse:
      properties:
        peer:
          $ref: '#/components/schemas/AdminIssuePeerPeerOut'
        config_awg:
          $ref: '#/components/schemas/ConfigEntryOut'
        config_wg_obf:
          $ref: '#/components/schemas/ConfigEntryOut'
        config_wg:
          $ref: '#/components/schemas/ConfigEntryOut'
        request_id:
          type: string
          title: Request Id
        peer_created:
          type: boolean
          title: Peer Created
          default: false
      type: object
      required:
      - peer
      - config_awg
      - config_wg_obf
      - config_wg
      - request_id
      title: AdminIssuePeerResponse
...
    AgentV1PeerOut:
      properties:
        public_key:
          type: string
          title: Public Key
        allowed_ips:
          type: string
          title: Allowed Ips
          default: ''
        last_handshake_age_sec:
          anyOf:
          - type: integer
          - type: 'null'
          title: Last Handshake Age Sec
        rx_bytes:
          type: integer
          title: Rx Bytes
          default: 0
        tx_bytes:
          type: integer
          title: Tx Bytes
          default: 0
      type: object
      required:
      - public_key
      title: AgentV1PeerOut
      description: Single peer in agent v1 peers list (from heartbeat or snapshot).
...
    ConfigEntryOut:
      properties:
        download_url:
          type: string
          title: Download Url
        qr_payload:
          type: string
          title: Qr Payload
      type: object
      required:
      - download_url
      - qr_payload
      title: ConfigEntryOut
      description: Download URL and QR payload for a single config format.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IssueResponse:
      properties:
        device_id:
          type: string
          title: Device Id
        issued_at:
          type: string
          format: date-time
          title: Issued At
        config:
          anyOf:
          - type: string
          - type: 'null'
          title: Config
        config_awg:
          anyOf:
          - type: string
          - type: 'null'
          title: Config Awg
        config_wg_obf:
          anyOf:
          - type: string
          - type: 'null'
          title: Config Wg Obf
        config_wg:
          anyOf:
          - type: string
          - type: 'null'
          title: Config Wg
        server_id:
          type: string
          title: Server Id
        subscription_id:
          type: string
          title: Subscription Id
        node_mode:
          type: string
          title: Node Mode
          default: mock
        peer_created:
          type: boolean
          title: Peer Created
          default: false
      type: object
      required:
      - device_id
      - issued_at
      - config
      - server_id
      - subscription_id
      title: IssueResponse
...
    MigrateBody:
      properties:
        target_node_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Node Id
      type: object
      title: MigrateBody
      description: Optional target node for migration; omit to use load balancer.
...
    PeerListItemOut:
      properties:
        peer_id:
          type: string
          title: Peer Id
        node_id:
          type: string
          title: Node Id
        user_id:
          type: integer
          title: User Id
        subscription_id:
          type: string
          title: Subscription Id
        public_key:
          type: string
          title: Public Key
        client_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Name
        status:
          type: string
          title: Status
        issued_at:
          type: string
          format: date-time
          title: Issued At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
      type: object
      required:
      - peer_id
      - node_id
      - user_id
      - subscription_id
      - public_key
      - client_name
      - status
      - issued_at
      - revoked_at
      title: PeerListItemOut
      description: Single peer in list (device as peer).
    PeerListOut:
      properties:
        peers:
          items:
            $ref: '#/components/schemas/PeerListItemOut'
          type: array
          title: Peers
        total:
          type: integer
          title: Total
      type: object
      required:
      - peers
      - total
      title: PeerListOut
      description: Peers list response.
    PeerOut:
      properties:
        public_key:
          type: string
          title: Public Key
        peer_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Peer Id
        allowed_ips:
          anyOf:
          - type: string
          - type: 'null'
          title: Allowed Ips
        last_handshake_ts:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Handshake Ts
        rx_bytes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Rx Bytes
        tx_bytes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Tx Bytes
        traffic_bytes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Traffic Bytes
        status:
          type: string
          title: Status
          default: unknown
      type: object
      required:
      - public_key
      title: PeerOut
      description: Node peer. peer_id present when mapped to a Device (for admin rotate/revoke).
...
    ServerPeersOut:
      properties:
        peers:
          items:
            $ref: '#/components/schemas/PeerOut'
          type: array
          title: Peers
        total:
          type: integer
          title: Total
        node_reachable:
          type: boolean
          title: Node Reachable
          default: true
      type: object
      required:
      - peers
      - total
      title: ServerPeersOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WgPeerCreateBody:
      properties:
        user_id:
          type: integer
          title: User Id
        subscription_id:
          type: string
          title: Subscription Id
        device_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Name
        server_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Server Id
      type: object
      required:
      - user_id
      - subscription_id
      title: WgPeerCreateBody
      description: Body for POST /wg/peer (provision peer = issue device).
    ErrorInfo:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
      required:
      - code
      - message
    ErrorMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
        request_id:
          type: string
      required:
      - timestamp
      - code
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          $ref: '#/components/schemas/ErrorInfo'
        meta:
          $ref: '#/components/schemas/ErrorMeta'
      required:
      - success
      - data
      - error
      - meta
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Bot API key (X-API-Key header).
    AgentToken:
      type: apiKey
      in: header
      name: X-Agent-Token
      description: Node-agent shared token. mTLS required at reverse-proxy.
  responses:
    ErrorResponse:
      description: Unified error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'